Refactor Apple Music provider into modular structure
